Refactor- move entity interfaces
