+++ src/entities/Genre.ts
@@ -1,0 +1,7 @@
+
+export interface Genre {
+    id: number;
+    name: string;
+    slug: string;
+    image_background: string;
+}
