+++ src/hooks/useGenres.ts
@@ -2,15 +2,9 @@ import { useQuery } from "@tanstack/react-query";
 import genres from "../data/genres";
 import APIClient from "../services/api-client";
 import ms from "ms";
+import { Genre } from "../entities/Genre";
 
 const apiClient = new APIClient<Genre>('/genres');
-
-export interface Genre{
-    id: number;
-    name: string;
-    slug: string;
-    image_background: string;
-}
 
 const useGenres = () => {
     return useQuery({
