+++ src/hooks/usePlatforms.ts
@@ -2,14 +2,9 @@ import { useQuery } from "@tanstack/react-query";
 import platforms from "../data/platforms";
 import APIClient from "../services/api-client";
 import ms from "ms";
+import { Platform } from "../entities/Platform";
 
 const apiClient = new APIClient<Platform>('/platforms/lists/parents');
-
-export interface Platform {
-    id: number;
-    name: string;
-    slug: string;
-}
 
 const usePlatforms = () => {
     return useQuery({
